vscode module move does bad job, sad
