+++ core/test/config/config-helpers-test.js
@@ -25,7 +25,7 @@ import BaseGatherer from '../../gather/base-gatherer.js';
 import ImageElementsGatherer from '../../gather/gatherers/image-elements.js';
 import UserTimingsAudit from '../../audits/user-timings.js';
 import {LH_ROOT} from '../../../root.js';
-import {getModuleDirectory} from '../../../esm-utils.js';
+import {getModuleDirectory} from '../../../shared/esm-utils.js';
 
 const require = createRequire(import.meta.url);
 const moduleDir = getModuleDirectory(import.meta);
