+++ core/test/lib/i18n/i18n-test.js
@@ -10,7 +10,7 @@ import log from 'lighthouse-logger';
 import jestMock from 'jest-mock';
 
 import * as i18n from '../../../lib/i18n/i18n.js';
-import {getModuleDirectory} from '../../../../esm-utils.js';
+import {getModuleDirectory} from '../../../../shared/esm-utils.js';
 
 const moduleDir = getModuleDirectory(import.meta);
